requirements.txt
flask
pandas
scikit-learn
psycopg2-binary
joblib
openpyxl
werkzeug
